T/M hoofdstuk 17

+++ src/main/java/nl/miwnn/ch19/paul/skirental/controller/CopyController.java
@@ -3,8 +3,7 @@ package nl.miwnn.ch19.paul.skirental.controller;
 import nl.miwnn.ch19.paul.skirental.model.Copy;
 import nl.miwnn.ch19.paul.skirental.repository.CopyRepository;
 import nl.miwnn.ch19.paul.skirental.repository.SkiRepository;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import nl.miwnn.ch19.paul.skirental.repository.SnowboardRepository;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -13,22 +12,20 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.util.Optional;
 
-/**
- * @author Paul Rademaker
- * Handle al requests regarding copies
- */
-
 @Controller
-@RequestMapping("/copies/")
+@RequestMapping("/copies")
 public class CopyController {
 
-    private static final Logger log = LoggerFactory.getLogger(CopyController.class);
     private final CopyRepository copyRepository;
     private final SkiRepository skiRepository;
+    private final SnowboardRepository snowboardRepository;
 
-    public CopyController(CopyRepository copyRepository, SkiRepository skiRepository) {
+    public CopyController(CopyRepository copyRepository,
+                          SkiRepository skiRepository,
+                          SnowboardRepository snowboardRepository) {
         this.copyRepository = copyRepository;
         this.skiRepository = skiRepository;
+        this.snowboardRepository = snowboardRepository;
     }
 
     @PostMapping("/borrow/{id}")
@@ -39,12 +36,10 @@ public class CopyController {
             Copy copy = optionalCopy.get();
             copy.setAvailable(false);
             copyRepository.save(copy);
-            log.info("Copy {} uitgeleend", id);
-            redirectAttributes.addFlashAttribute("successMessage", "Ski uitgeleend!");
-            return "redirect:/ski/" + copy.getSki().getId();
+            redirectAttributes.addFlashAttribute("successMessage", "Item uitgeleend!");
+            return getRedirectUrl(copy);
         }
-
-        return "redirect:/ski/all";
+        return "redirect:/";
     }
 
     @PostMapping("/return/{id}")
@@ -55,34 +50,51 @@ public class CopyController {
             Copy copy = optionalCopy.get();
             copy.setAvailable(true);
             copyRepository.save(copy);
-            log.info("Copy {} teruggebracht", id);
-            redirectAttributes.addFlashAttribute("successMessage", "Ski teruggebracht!");
-            return "redirect:/ski/" + copy.getSki().getId();
+            redirectAttributes.addFlashAttribute("successMessage", "Item teruggebracht!");
+            return getRedirectUrl(copy);
         }
-
-        return "redirect:/ski/all";
+        return "redirect:/";
     }
 
     @PostMapping("/add/{skiId}")
-    public String addCopy(@PathVariable Long skiId, RedirectAttributes redirectAttributes) {
+    public String addSkiCopy(@PathVariable Long skiId, RedirectAttributes redirectAttributes) {
         skiRepository.findById(skiId).ifPresent(ski -> {
-            Copy copy = new Copy(ski);
-            copyRepository.save(copy);
-            log.info("Nieuw exemplaar toegevoegd aan ski {}", skiId);
+            copyRepository.save(new Copy(ski));
         });
-        redirectAttributes.addFlashAttribute("successMessage", "Exemplaar toegevoegd!");
+        redirectAttributes.addFlashAttribute("successMessage", "Ski exemplaar toegevoegd!");
         return "redirect:/ski/" + skiId;
+    }
+
+    // DIT IS DE NIEUWE ROUTE VOOR SNOWBOARDS
+    @PostMapping("/add-snowboard/{snowboardId}")
+    public String addSnowboardCopy(@PathVariable Long snowboardId, RedirectAttributes redirectAttributes) {
+        snowboardRepository.findById(snowboardId).ifPresent(snowboard -> {
+            copyRepository.save(new Copy(snowboard));
+        });
+        redirectAttributes.addFlashAttribute("successMessage", "Snowboard exemplaar toegevoegd!");
+        return "redirect:/snowboard/" + snowboardId;
     }
 
     @PostMapping("/delete/{id}")
     public String deleteCopy(@PathVariable Long id, RedirectAttributes redirectAttributes) {
-        copyRepository.findById(id).ifPresent(copy -> {
-            Long skiId = copy.getSki().getId();
+        Optional<Copy> optionalCopy = copyRepository.findById(id);
+        if (optionalCopy.isPresent()) {
+            Copy copy = optionalCopy.get();
+            String redirectUrl = getRedirectUrl(copy);
             copyRepository.deleteById(id);
-            log.info("Exemplaar {} verwijderd", id);
             redirectAttributes.addFlashAttribute("successMessage", "Exemplaar verwijderd!");
-        });
+            return redirectUrl;
+        }
+        return "redirect:/";
+    }
 
-        return "redirect:/ski/all";
+    // Handige hulp-methode om te bepalen of we terug moeten naar de ski- of snowboard-pagina
+    private String getRedirectUrl(Copy copy) {
+        if (copy.getSki() != null) {
+            return "redirect:/ski/" + copy.getSki().getId();
+        } else if (copy.getSnowboard() != null) {
+            return "redirect:/snowboard/" + copy.getSnowboard().getId();
+        }
+        return "redirect:/";
     }
 }
